LatticeBatch and Friends Handle Buffer Sizes
This change makes it so that the types of the buffers are determined
by the schema used to create the batch.

The size of each buffer is tracked in the batch and used by the
writer. On read, the number of values in the column is looked up by
the reader.

It assumes a 1:1 between LatticeRowBatch:Stripe. This is obviously
false but this a working step towards that.